General bindings formatting
